+++ source/core/tests/helpers/schemas/bronze_submitted_transactions_value_schema.py
@@ -55,3 +55,46 @@ bronze_submitted_transactions_value_schema = StructType(
         StructField(BronzeSubmittedTransactionsColumnNames.partition, IntegerType(), True),
     ]
 )
+
+
+# For test purpose, enums are integers for when we deserialize the protobuf message
+deserialized_bronze_submitted_transactions_value_schema = StructType(
+    [
+        StructField(ValueColumnNames.version, StringType(), True),
+        StructField(ValueColumnNames.orchestration_instance_id, StringType(), True),
+        StructField(ValueColumnNames.orchestration_type, IntegerType(), True),
+        StructField(ValueColumnNames.metering_point_id, StringType(), True),
+        StructField(ValueColumnNames.transaction_id, StringType(), True),
+        StructField(ValueColumnNames.transaction_creation_datetime, TimestampType(), True),
+        StructField(ValueColumnNames.metering_point_type, IntegerType(), True),
+        StructField(ValueColumnNames.unit, IntegerType(), True),
+        StructField(ValueColumnNames.resolution, IntegerType(), True),
+        StructField(ValueColumnNames.start_datetime, TimestampType(), True),
+        StructField(ValueColumnNames.end_datetime, TimestampType(), True),
+        StructField(
+            ValueColumnNames.points,
+            ArrayType(
+                StructType(
+                    [
+                        StructField(ValueColumnNames.Points.position, IntegerType(), True),
+                        StructField(
+                            ValueColumnNames.Points.quantity,
+                            StructType(
+                                [
+                                    StructField("units", LongType(), False),
+                                    StructField("nanos", IntegerType(), False),
+                                ]
+                            ),
+                            True,
+                        ),
+                        StructField(ValueColumnNames.Points.quality, IntegerType(), True),
+                    ]
+                ),
+                True,
+            ),
+            True,
+        ),
+        StructField(BronzeSubmittedTransactionsColumnNames.key, BinaryType(), True),
+        StructField(BronzeSubmittedTransactionsColumnNames.partition, IntegerType(), True),
+    ]
+)