+++ source/dotnet/Measurements.Abstractions/Api/Models/AggregatedMeasurementDto.cs
@@ -1,3 +1,0 @@
-﻿namespace Energinet.DataHub.Measurements.Abstractions.Api.Models;
-
-public sealed record MeasurementAggregationDto(DateOnly Date, decimal Quantity, Quality Quality, Unit Unit, bool MissingValues, bool ContainsUpdatedValues);