+++ source/dotnet/Measurements.Application/Requests/GetAggregatedByDateRequest.cs
@@ -1,0 +1,5 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Energinet.DataHub.Measurements.Application.Requests;
+
+public record GetAggregatedByDateRequest(string MeteringPointId, [Range(-9998, 9999)] int Year, [Range(1, 12)] int Month);