+++ source/dotnet/Measurements.WebApi/Controllers/MeasurementsController.cs
@@ -8,36 +8,14 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Energinet.DataHub.Measurements.WebApi.Controllers;
 
-[ApiVersion(1.0, Deprecated = true)]
-[ApiVersion(2.0)]
 [ApiController]
 [Authorize]
-public class MeasurementsController(IMeasurementsHandler measurementsHandler, ILogger<MeasurementsController> logger)
+[ApiVersion(2.0)]
+[Route("v{v:apiVersion}/measurements")]
+public class MeasurementsController(
+    IMeasurementsHandler measurementsHandler, ILogger<MeasurementsController> logger, IJsonSerializer jsonSerializer)
     : ControllerBase
 {
-    [MapToApiVersion(1.0)]
-    [HttpGet("forPeriod")]
-    public async Task<IActionResult> GetByPeriodAsyncV1([FromQuery] GetByPeriodRequest request)
-    {
-        try
-        {
-            var measurement = await measurementsHandler.GetByPeriodAsyncV1(request);
-            var result = new JsonSerializer().Serialize(measurement);
-
-            return Ok(result);
-        }
-        catch (MeasurementsNotFoundDuringPeriodException e)
-        {
-            return NotFound(e.Message);
-        }
-        catch (Exception exception)
-        {
-            logger.LogError(exception, "Could not get requested measurement");
-
-            return StatusCode(StatusCodes.Status500InternalServerError, exception.Message);
-        }
-    }
-
     [MapToApiVersion(2.0)]
     [HttpGet("forPeriod")]
     public async Task<IActionResult> GetByPeriodAsync([FromQuery] GetByPeriodRequest request)
@@ -45,7 +23,7 @@ public class MeasurementsController(IMeasurementsHandler measurementsHandler, IL
         try
         {
             var measurement = await measurementsHandler.GetByPeriodAsync(request);
-            var result = new JsonSerializer().Serialize(measurement);
+            var result = jsonSerializer.Serialize(measurement);
 
             return Ok(result);
         }
@@ -55,27 +33,55 @@ public class MeasurementsController(IMeasurementsHandler measurementsHandler, IL
         }
         catch (Exception exception)
         {
-            logger.LogError(exception, "Could not get requested measurement");
+            logger.LogError(exception, "Could not get requested measurements");
 
             return StatusCode(StatusCodes.Status500InternalServerError, exception.Message);
         }
     }
 
-    [MapToApiVersion(1.0)]
     [MapToApiVersion(2.0)]
     [HttpGet("aggregatedByMonth")]
-    public async Task<IActionResult> GetAggregatedByMonthAsync([FromQuery] GetAggregatedByMonthRequest request)
+    public async Task<IActionResult> GetAggregatedByDateAsync([FromQuery] GetAggregatedByDateRequest request)
     {
         try
         {
-            var aggregatedByMonth = await measurementsHandler.GetAggregatedByMonthAsync(request);
-            var result = new JsonSerializer().Serialize(aggregatedByMonth);
+            var aggregatedByMonth = await measurementsHandler.GetAggregatedByDateAsync(request);
+            var result = jsonSerializer.Serialize(aggregatedByMonth);
 
             return Ok(result);
         }
         catch (MeasurementsNotFoundDuringPeriodException e)
         {
             return NotFound(e.Message);
+        }
+        catch (Exception exception)
+        {
+            logger.LogError(exception, "Could not get requested measurements");
+
+            return StatusCode(StatusCodes.Status500InternalServerError, exception.Message);
+        }
+    }
+
+    [MapToApiVersion(2.0)]
+    [HttpGet("aggregatedByYear")]
+    public async Task<IActionResult> GetAggregatedByMonthAsync([FromQuery] GetAggregatedByMonthRequest request)
+    {
+        try
+        {
+            var aggregatedByYear = await measurementsHandler.GetAggregatedByMonthAsync(request);
+            var result = jsonSerializer.Serialize(aggregatedByYear);
+
+            return Ok(result);
+        }
+        catch (MeasurementsNotFoundDuringPeriodException e)
+        {
+            return NotFound(e.Message);
+        }
+        catch (Exception exception)
+        {
+            logger.LogError(exception, "Could not get requested measurements");
+
+            return StatusCode(StatusCodes.Status500InternalServerError, exception.Message);
         }
     }