+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/infrastructure/__init__.py
@@ -1,17 +1,13 @@
-from geh_calculated_measurements.common.infrastructure.calculated_measurements.database_definitions import (
-    CalculatedMeasurementsDatabaseDefinition,
-    CalculatedMeasurementsInternalDatabaseDefinition,
-)
 from geh_calculated_measurements.common.infrastructure.calculated_measurements.repository import (
     Repository as CalculatedMeasurementsRepository,
-)
-from geh_calculated_measurements.common.infrastructure.current_measurements.current_meaurements_table import (
-    CurrentMeasurementsTable,
 )
 from geh_calculated_measurements.common.infrastructure.current_measurements.repository import (
     Repository as CurrentMeasurementsRepository,
 )
-from geh_calculated_measurements.common.infrastructure.current_measurements.table import Table
+from geh_calculated_measurements.common.infrastructure.database_definitions import (
+    CalculatedMeasurementsDatabaseDefinition,
+    CalculatedMeasurementsInternalDatabaseDefinition,
+)
 from geh_calculated_measurements.common.infrastructure.spark_initializor import initialize_spark
 
 __all__ = [
@@ -19,7 +15,6 @@ __all__ = [
     "CalculatedMeasurementsDatabaseDefinition",
     "CalculatedMeasurementsRepository",
     "CurrentMeasurementsRepository",
-    "CurrentMeasurementsTable",
     "initialize_spark",
     "Table",
 ]