+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/infrastructure/calculated_measurements/repository.py
@@ -1,12 +1,8 @@
 from pyspark.sql import SparkSession
 
 from geh_calculated_measurements.common.application.model import CalculatedMeasurementsInternal
-from geh_calculated_measurements.common.domain.model.current_measurements import CurrentMeasurements
-from geh_calculated_measurements.common.infrastructure.calculated_measurements.database_definitions import (
+from geh_calculated_measurements.common.infrastructure.database_definitions import (
     CalculatedMeasurementsInternalDatabaseDefinition,
-)
-from geh_calculated_measurements.common.infrastructure.current_measurements.database_definitions import (
-    MeasurementsGoldDatabaseDefinition,
 )
 
 
@@ -29,9 +25,3 @@ class Repository:
         database_name = CalculatedMeasurementsInternalDatabaseDefinition.DATABASE_NAME
         table_name = CalculatedMeasurementsInternalDatabaseDefinition.MEASUREMENTS_TABLE_NAME
         df.write.format("delta").mode("append").saveAsTable(self._get_full_table_path(database_name, table_name))
-
-    def read_current_measurements(self) -> CurrentMeasurements:
-        database_name = MeasurementsGoldDatabaseDefinition.DATABASE_NAME
-        table_name = MeasurementsGoldDatabaseDefinition.CURRENT_MEASUREMENTS
-        df = self._spark.table(self._get_full_table_path(database_name, table_name))
-        return CurrentMeasurements(df)