+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/infrastructure/current_measurements/database_definitions.py
@@ -1,4 +1,0 @@
-# TODO JMG: Use values from geh_common.data_products
-class MeasurementsGoldDatabaseDefinition:
-    DATABASE_NAME = "measurements_gold"
-    CURRENT_MEASUREMENTS = "current_v1"