+++ source/geh_calculated_measurements/tests/subsystem_tests/missing_measurements_log/test_jobs.py
@@ -3,9 +3,10 @@ from datetime import timedelta
 
 import pytest
 
-from geh_calculated_measurements.testing.utilities.job_tester import JobTest, JobTestFixture
+from geh_calculated_measurements.common.infrastructure import CalculatedMeasurementsInternalDatabaseDefinition
+from geh_calculated_measurements.testing import JobTest, JobTestFixture
 from tests.subsystem_tests.environment_configuration import EnvironmentConfiguration
-from tests.subsystem_tests.missing_measurements_log.seed_table import PERIOD_START, seed_table
+from tests.subsystem_tests.missing_measurements_log.seed_table import PERIOD_START, delete_seeded_data, seed_table
 
 period_start = PERIOD_START
 period_end = period_start + timedelta(days=2)
@@ -18,22 +19,27 @@ job_parameters = {
 
 class TestMissingMeasurementsLog(JobTest):
     @pytest.fixture(scope="class")
-    def fixture(self, electricity_market_data_products_created_as_tables):
+    def fixture(self):
         config = EnvironmentConfiguration()
         base_job_fixture = JobTestFixture(
             environment_configuration=config,
             job_name="MissingMeasurementsLog",
             job_parameters=job_parameters,
         )
+        delete_seeded_data(base_job_fixture)
 
         seed_table(base_job_fixture)
 
+        yield base_job_fixture
+
+        delete_seeded_data(base_job_fixture)
+
         return base_job_fixture
 
-    @pytest.mark.skip(reason="Skipped due to issues with the telemetry data not available in the logs.")
-    def test__and_then_job_telemetry_is_created(self, job_fixture) -> None:
-        pass
-
-    @pytest.mark.skip(reason="This test is temporary skipped because the storing implementation is not yet made.")
-    def test__and_then_data_is_written_to_delta(self, job_fixture) -> None:
-        pass
+    def test__and_then_data_is_written_to_delta(self, fixture) -> None:
+        """
+        Test that data is written to the delta table.
+        """
+        database_name = CalculatedMeasurementsInternalDatabaseDefinition.DATABASE_NAME
+        table_name = CalculatedMeasurementsInternalDatabaseDefinition.MISSING_MEASUREMENTS_LOG_TABLE_NAME
+        self.assert_data_written_to_delta(fixture, database_name, table_name)